refactor(#54): consolidate configs
